update attendance and lecture repo

+++ app/src/main/java/lk/lnbti/iampresent/repo/LectureRepo.kt
@@ -2,13 +2,23 @@ package lk.lnbti.iampresent.repo
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import lk.lnbti.iampresent.dao.LectureDao
 import lk.lnbti.iampresent.data.Lecture
 import lk.lnbti.iampresent.data.Result
-import lk.lnbti.iampresent.dao.LectureDao
 import java.io.IOException
 import javax.inject.Inject
 
+/**
+ * Repository class for handling operations related to lectures.
+ *
+ * @property lectureDao Data Access Object for lectures.
+ */
 class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
+    /**
+     * Retrieves a list of all lectures.
+     *
+     * @return Result object containing either a list of lectures on success or an error message on failure.
+     */
     suspend fun findLectureList(): Result<List<Lecture>> {
         return withContext(Dispatchers.IO) {
             return@withContext try {
@@ -23,6 +33,12 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
             }
         }
     }
+
+    /**
+     * Retrieves a list of lectures scheduled for today.
+     *
+     * @return Result object containing either a list of lectures on success or an error message on failure.
+     */
     suspend fun findTodaysLectureList(): Result<List<Lecture>> {
         return withContext(Dispatchers.IO) {
             return@withContext try {
@@ -37,12 +53,19 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
             }
         }
     }
+
+    /**
+     * Saves a new lecture.
+     *
+     * @param lecture The lecture to be saved.
+     * @return Result object containing either the saved lecture on success or an error message on failure.
+     */
     suspend fun saveLecture(lecture: Lecture): Result<Lecture?> {
         return withContext(Dispatchers.IO) {
             return@withContext try {
                 val response = lectureDao.saveLecture(lecture)
                 if (response.isSuccessful) {
-                    Result.Success(response.body()?:null)
+                    Result.Success(response.body() ?: null)
                 } else {
                     Result.Error("Failed to save lecture")
                 }
@@ -52,7 +75,13 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
         }
     }
 
-    suspend fun deleteLecture(lectureId: Int): Result<Lecture?> {
+    /**
+     * Deactivates a lecture by its ID.
+     *
+     * @param lectureId The ID of the lecture to be deactivated.
+     * @return Result object containing either the deactivated lecture on success or an error message on failure.
+     */
+    suspend fun deleteLecture(lectureId: Long): Result<Lecture?> {
         return withContext(Dispatchers.IO) {
             return@withContext try {
                 val response = lectureDao.deleteLectureById(lectureId)
@@ -66,18 +95,25 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
             }
         }
     }
+
+    /**
+     * Retrieves a lecture by its ID.
+     *
+     * @param lectureId The ID of the lecture to be retrieved.
+     * @return The retrieved lecture or null if not found.
+     */
     suspend fun findLectureById(lectureId: String): Lecture? {
         return withContext(Dispatchers.IO) {
             return@withContext try {
                 getLecture(lectureId = lectureId)
-            }catch (e:Exception){
+            } catch (e: Exception) {
                 null
             }
         }
     }
 
     private suspend fun getLecture(lectureId: String): Lecture? {
-        val response = lectureDao.findLectureById(lectureId.toInt())
+        val response = lectureDao.findLectureById(lectureId.toLong())
         lateinit var lecture: Lecture
         if (response.isSuccessful) {
             lecture = response.body()!!
@@ -85,13 +121,19 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
         return lecture
     }
 
-    suspend fun openLectureForAttendance(lectureId: Int): Lecture? {
+    /**
+     * Opens a lecture for attendance by updating its status.
+     *
+     * @param lectureId The ID of the lecture to be opened for attendance.
+     * @return The updated lecture or null on failure.
+     */
+    suspend fun openLectureForAttendance(lectureId: Long): Lecture? {
         return withContext(Dispatchers.IO) {
             return@withContext updateLectureStatusToOpen(lectureId = lectureId)
         }
     }
 
-    private suspend fun updateLectureStatusToOpen(lectureId: Int): Lecture? {
+    private suspend fun updateLectureStatusToOpen(lectureId: Long): Lecture? {
         val response = lectureDao.openLectureForAttendance(lectureId)
         lateinit var lecture: Lecture
         if (response.isSuccessful) {
@@ -100,13 +142,19 @@ class LectureRepo @Inject constructor(private val lectureDao: LectureDao) {
         return lecture
     }
 
-    suspend fun closeLectureForAttendance(lectureId: Int): Lecture? {
+    /**
+     * Closes a lecture for attendance by updating its status.
+     *
+     * @param lectureId The ID of the lecture to be closed for attendance.
+     * @return The updated lecture or null on failure.
+     */
+    suspend fun closeLectureForAttendance(lectureId: Long): Lecture? {
         return withContext(Dispatchers.IO) {
             return@withContext updateLectureStatusToClose(lectureId = lectureId)
         }
     }
 
-    private suspend fun updateLectureStatusToClose(lectureId: Int): Lecture? {
+    private suspend fun updateLectureStatusToClose(lectureId: Long): Lecture? {
         val response = lectureDao.closeLectureForAttendance(lectureId)
         lateinit var lecture: Lecture
         if (response.isSuccessful) {
